% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FrequencyResult-class.R
\docType{class}
\name{FrequencyResult-class}
\alias{FrequencyResult-class}
\title{Title}
\description{
Title
}
\section{Slots}{

\describe{
\item{\code{data}}{array.}

\item{\code{frequency}}{vector.}

\item{\code{wavenumber}}{vector.}
}}

