% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{imshow}
\alias{imshow}
\title{Title}
\usage{
imshow(x = seq(0, 1, length.out = nrow(z)), y = seq(0, 1, length.out =
  ncol(z)), z, type = "contour", interactive = FALSE, component = NA, ...)
}
\arguments{
\item{x}{}

\item{y}{}

\item{z}{}

\item{type}{}

\item{interactive}{}

\item{component}{}

\item{...}{}
}
\description{
Title
}

