\name{OscillationsApp-package}
\alias{OscillationsApp-package}
\alias{OscillationsApp}
\docType{package}
\title{
\packageTitle{OscillationsApp}
}
\description{
\packageDescription{OscillationsApp}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{OscillationsApp}
\packageIndices{OscillationsApp}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{OscillationsApp}

Maintainer: \packageMaintainer{OscillationsApp}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
