% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src.R
\name{trimData}
\alias{trimData}
\title{Title}
\usage{
trimData(object, remove.negative = TRUE, uniform = FALSE, start.time = NA,
  end.time = NA)
}
\arguments{
\item{object}{}

\item{remove.negative}{}

\item{uniform}{}

\item{start.time}{}

\item{end.time}{}
}
\description{
Title
}

