% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wavelet.R
\name{wavelet}
\alias{wavelet}
\title{Title}
\usage{
wavelet(object, noctave = 8, nvoice = 8, fc = 1, ...)
}
\arguments{
\item{object}{}

\item{noctave}{}

\item{nvoice}{}

\item{fc}{}

\item{...}{}
}
\description{
Title
}

