% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WaveletResult-class.R
\docType{class}
\name{WaveletResult-class}
\alias{WaveletResult-class}
\title{Title}
\description{
Title
}
\section{Slots}{

\describe{
\item{\code{data}}{array.}

\item{\code{wavenumber}}{vector.}

\item{\code{time}}{vector.}

\item{\code{frequency}}{vector.}
}}

